rename use selected element settings
