+++ packages/packages/core/editor-components/src/hooks/use-sanitize-overridable-props.ts
@@ -15,8 +15,6 @@ export function useSanitizeOverridableProps(
 		return undefined;
 	}
 
-	// return overridableProps;
-
 	if ( isSanitized ) {
 		return overridableProps;
 	}
